requirements.txt
flask>=2.3
dash>=2.14
dash-bootstrap-components>=1.5
plotly>=5.18
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
pdfplumber>=0.10
requests>=2.31
beautifulsoup4>=4.12
Pillow>=10.0
python-dotenv>=1.0
reportlab>=4.0
pymongo[srv]>=4.6
gunicorn>=21.2
pytest>=7.4